Move SawyerCoding from util into core, and into its own namespace
