+++ src/openrct2/Game.cpp
@@ -30,6 +30,7 @@
 #include "core/FileScanner.h"
 #include "core/Money.hpp"
 #include "core/Path.hpp"
+#include "core/SawyerCoding.h"
 #include "entity/EntityRegistry.h"
 #include "entity/PatrolArea.h"
 #include "entity/Peep.h"
@@ -59,7 +60,6 @@
 #include "scripting/ScriptEngine.h"
 #include "ui/UiContext.h"
 #include "ui/WindowManager.h"
-#include "util/SawyerCoding.h"
 #include "util/Util.h"
 #include "windows/Intent.h"
 #include "world/Banner.h"
