+++ src/openrct2/rct2/S6Importer.cpp
@@ -21,6 +21,7 @@
 #include "../core/Numerics.hpp"
 #include "../core/Path.hpp"
 #include "../core/Random.hpp"
+#include "../core/SawyerCoding.h"
 #include "../core/String.hpp"
 #include "../entity/Balloon.h"
 #include "../entity/Duck.h"
@@ -68,7 +69,6 @@
 #include "../scenario/Scenario.h"
 #include "../scenario/ScenarioRepository.h"
 #include "../scenario/ScenarioSources.h"
-#include "../util/SawyerCoding.h"
 #include "../util/Util.h"
 #include "../world/Climate.h"
 #include "../world/Entrance.h"
