+++ src/openrct2/ride/Track.cpp
@@ -15,13 +15,13 @@
 #include "../GameState.h"
 #include "../audio/audio.h"
 #include "../config/Config.h"
+#include "../core/SawyerCoding.h"
 #include "../interface/Viewport.h"
 #include "../management/Finance.h"
 #include "../network/network.h"
 #include "../platform/Platform.h"
 #include "../rct1/RCT1.h"
 #include "../ride/RideColour.h"
-#include "../util/SawyerCoding.h"
 #include "../world/Footpath.h"
 #include "../world/Map.h"
 #include "../world/MapAnimation.h"
