+++ src/openrct2/ride/TrackDesign.cpp
@@ -34,6 +34,7 @@
 #include "../core/DataSerialiser.h"
 #include "../core/File.h"
 #include "../core/Numerics.hpp"
+#include "../core/SawyerCoding.h"
 #include "../core/String.hpp"
 #include "../core/UnitConversion.h"
 #include "../drawing/X8DrawingEngine.h"
@@ -52,7 +53,6 @@
 #include "../object/StationObject.h"
 #include "../rct2/RCT2.h"
 #include "../ride/RideConstruction.h"
-#include "../util/SawyerCoding.h"
 #include "../util/Util.h"
 #include "../world/Footpath.h"
 #include "../world/Park.h"
