+++ src/openrct2/ride/TrackDesignSave.cpp
@@ -11,6 +11,7 @@
 #include "../Diagnostic.h"
 #include "../Game.h"
 #include "../audio/audio.h"
+#include "../core/SawyerCoding.h"
 #include "../interface/Viewport.h"
 #include "../localisation/StringIds.h"
 #include "../object/FootpathObject.h"
@@ -20,7 +21,6 @@
 #include "../object/ObjectList.h"
 #include "../object/ObjectManager.h"
 #include "../rct2/RCT2.h"
-#include "../util/SawyerCoding.h"
 #include "../windows/Intent.h"
 #include "../world/Footpath.h"
 #include "../world/Scenery.h"
