+++ src/openrct2/scenario/Scenario.cpp
@@ -24,6 +24,7 @@
 #include "../core/Guard.hpp"
 #include "../core/Path.hpp"
 #include "../core/Random.hpp"
+#include "../core/SawyerCoding.h"
 #include "../core/UnitConversion.h"
 #include "../entity/Duck.h"
 #include "../entity/Guest.h"
@@ -47,7 +48,6 @@
 #include "../rct12/RCT12.h"
 #include "../ride/Ride.h"
 #include "../ride/Track.h"
-#include "../util/SawyerCoding.h"
 #include "../util/Util.h"
 #include "../windows/Intent.h"
 #include "../world/Climate.h"
